Base pyboy class
