+++ scripts/envrionments/pyboy/pokemon/map_constants.json
