+++ scripts/envrionments/pyboy/pokemon/pokemon_constants.json
