+++ scripts/envrionments/pyboy/pokemon/type_constants.json
